+++ src/API/Types.hs
@@ -848,6 +848,7 @@ data QualityReportAPI = QualityReportAPI
     , qraSuspiciousAmounts :: QualityCheckAPI
     , qraMissingMetadata :: QualityCheckAPI
     , qraFormulaConsistency :: QualityCheckAPI
+    , qraTruncatedNameCollisions :: QualityCheckAPI
     }
     deriving (Generic)
     deriving (ToJSON, FromJSON, ToSchema) via (Stripped QualityReportAPI)